QE-1471: Moved an import helper to the appropriate helper file
